[Build Speed] Remove DocumentInlines.h from all non-Inlines.h headers rdar://160166530 https://bugs.webkit.org/show_bug.cgi?id=298578
Reviewed by Chris Dumez.

Prior to this patch, there were 7 instances of including Document.h in non-Inlines.h headers, the
third largest such header in the WebCore project.

Many of the uses of DocumentInlines.h was to enable using MutationObserverOptions, so this enum
and OptionSet were pulled into its own (inexpensive) header so it could be used directly in Document.h.

Other uses included going from Document to Frame to Page in InspectorInstrumentation, but since the
entry points passing Page were non-inline, there was no benefit to inlining these functions.

DocumentFullscreen.h needed DocumentInlines.h to implement a private, inline nested class; the entire class
was moved into the implementation file. A private page() accessor was moved to the implementation file as well.

Other files included DocumentInlines.h but did not need to, including HTMLDocument.h and XMLDocument.h.

After this patch, no uses of Document.h in non-Inlines.h headers remain.

* Source/WebCore/Headers.cmake:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/bindings/js/JSCustomElementInterface.cpp:
* Source/WebCore/bindings/js/JSWindowProxy.cpp:
* Source/WebCore/dom/ChildListMutationScope.h:
* Source/WebCore/dom/Document.h:
(WebCore::Document::hasMutationObserversOfType const):
* Source/WebCore/dom/DocumentFullscreen.cpp:
(WebCore::DocumentFullscreen::page const):
* Source/WebCore/dom/DocumentFullscreen.h:
* Source/WebCore/dom/DocumentInlines.h:
(WebCore::Document::hasMutationObserversOfType const): Deleted.
* Source/WebCore/dom/MutationObserver.h:
* Source/WebCore/dom/MutationObserverInterestGroup.h:
* Source/WebCore/dom/MutationObserverOptions.h:
* Source/WebCore/dom/XMLDocument.h:
* Source/WebCore/html/HTMLDocument.h:
* Source/WebCore/html/canvas/WebGLRenderingContext.cpp:
* Source/WebCore/inspector/InspectorInstrumentation.cpp:
(WebCore::InspectorInstrumentation::instrumentingAgents):
* Source/WebCore/inspector/InspectorInstrumentation.h:
(WebCore::InspectorInstrumentation::didAddOrRemoveScrollbars):
(WebCore::InspectorInstrumentation::didReceiveThreadableLoaderResponse):
* Source/WebCore/loader/DocumentThreadableLoader.cpp:
(WebCore::DocumentThreadableLoader::didReceiveResponse):
* Source/WebCore/style/StyleExtractorConverter.h:
* Source/WebCore/xml/DOMParser.cpp:
* Source/WebKitLegacy/mac/DOM/DOMUtility.mm:

Canonical link: https://commits.webkit.org/299782@main
