+++ Source/WebCore/xml/DOMParser.cpp
@@ -20,6 +20,7 @@
 #include "DOMParser.h"
 
 #include "CommonAtomStrings.h"
+#include "DocumentInlines.h"
 #include "ExceptionOr.h"
 #include "HTMLDocument.h"
 #include "NodeInlines.h"
